+++ tests/linked_list_secret.h
@@ -1,12 +1,12 @@
 #ifndef LINKED_LIST_SECRET_H
 #define LINKED_LIST_SECRET_H
 
-#include <stdlib.h>
+#include <cstdint> // for uint64_t
 
 struct Node;
 
-struct Node* create_node(void* ptr, size_t amount, void** children);
+struct Node *create_node (void *ptr, uint64_t amount, void **children);
 
-void free_node(struct Node* current);
+void free_node (struct Node *current);
 
-#endif  // !LINKED_LIST_SECRET_H
+#endif // !LINKED_LIST_SECRET_H
